+++ wtforms/widgets/core.py
@@ -1,6 +1,6 @@
 from __future__ import unicode_literals
 
-from markupsafe import escape, Markup
+from markupsafe import Markup, escape
 
 from wtforms.compat import iteritems, text_type